Adding setup cleanup and removed some dependencies
