+++ test/test-cases/functional/saic/sai-api/test_sai_api_vnet_direction_lookup.py
