+++ test/test-cases/functional/saic/sai-api/test_sai_api_vnet_eni_addr.py
