+++ test/test-cases/functional/saic/sai-api/test_sai_api_vnet_out_route.py
@@ -1,0 +1,226 @@
+import json
+import time
+from pathlib import Path
+from pprint import pprint
+
+import pytest
+
+# Constants
+SWITCH_ID = 5
+
+
+class TestSaiVnetOutboundRoutingEntry:
+
+    def test_vnet_outbound_routing_entry_create(self, dpu):
+
+        # Create VNET
+        commands = [
+            {
+                "name": "vnet",
+                "op": "create",
+                "type": "SAI_OBJECT_TYPE_VNET",
+                "attributes": [
+                    "SAI_VNET_ATTR_VNI",
+                    "9000"
+                ]
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all(results), "SAI_OBJECT_TYPE_VNET Create error"
+        
+        # Create ENI
+        commands = [
+            {
+                "name": "eni_id",
+                "op": "create",
+                "type": "SAI_OBJECT_TYPE_ENI",
+                "attributes": [
+                    "SAI_ENI_ATTR_CPS",
+                    "10000",
+                    "SAI_ENI_ATTR_PPS",
+                    "100000",
+                    "SAI_ENI_ATTR_FLOWS",
+                    "100000",
+                    "SAI_ENI_ATTR_ADMIN_STATE",
+                    "True",
+                    "SAI_ENI_ATTR_VM_UNDERLAY_DIP",
+                    "10.10.9.10",
+                    "SAI_ENI_ATTR_VM_VNI",
+                    "9",
+                    "SAI_ENI_ATTR_VNET_ID",
+                    "$vnet",
+                    "SAI_ENI_ATTR_INBOUND_V4_STAGE1_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V4_STAGE2_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V4_STAGE3_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V4_STAGE4_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V4_STAGE5_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V6_STAGE1_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V6_STAGE2_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V6_STAGE3_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V6_STAGE4_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_INBOUND_V6_STAGE5_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V4_STAGE1_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V4_STAGE2_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V4_STAGE3_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V4_STAGE4_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V4_STAGE5_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V6_STAGE1_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V6_STAGE2_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V6_STAGE3_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V6_STAGE4_DASH_ACL_GROUP_ID",
+                    "0",
+                    "SAI_ENI_ATTR_OUTBOUND_V6_STAGE5_DASH_ACL_GROUP_ID",
+                    "0"
+                ]
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all(results), "SAI_OBJECT_TYPE_ENI Create error"
+        
+        commands = [
+            {
+                "name": "ore",
+                "op": "create",
+                "type": "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY",
+                "key": {
+                    "switch_id": "$SWITCH_ID",
+                    "eni_id": "$eni_id",
+                    "destination": "10.1.0.0/16"
+                },
+                "attributes": [
+                    "SAI_OUTBOUND_ROUTING_ENTRY_ATTR_ACTION", "SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET",
+                    "SAI_OUTBOUND_ROUTING_ENTRY_ATTR_DST_VNET_ID", "$vnet"
+                ]
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all(results), "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY Create error"
+
+    @pytest.mark.skip(reason="get and set not implemented, yet")
+    def test_vnet_outbound_routing_entry_get1(self, dpu):
+
+        commands = [
+            {
+                "name": "ore",
+                "op": "get",
+                "type": "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY",
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all( [result == 0 for result in results]), "SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET Get error"
+
+    @ pytest.mark.skip(reason="get and set not implemented, yet")
+    def test_vnet_outbound_routing_entry_set(self, dpu):
+
+        commands = [
+            {
+                "name": "ore",
+                "op": "create",
+                "type": "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY",
+                "key": {
+                    "switch_id": "$SWITCH_ID",
+                    "eni_id": "1",
+                    "destination": "10.1.1.0/16"
+                },
+                "attributes": [
+                    "SAI_OUTBOUND_ROUTING_ENTRY_ATTR_ACTION", "SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET",
+                    "SAI_OUTBOUND_ROUTING_ENTRY_ATTR_DST_VNET_ID", "$vnet"
+                ]
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+        
+        assert all( [result == 0 for result in results]), "SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET Set error"
+
+    @pytest.mark.skip(reason="get and set not implemented, yet")
+    def test_vnet_outbound_routing_entry_get2(self, dpu):
+
+        commands = [
+            {
+                "name": "ore",
+                "op": "get",
+                "type": "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY",
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all( [result == 0 for result in results]), "SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET Get error"
+
+    def test_vnet_outbound_routing_entry_remove(self, dpu):
+
+        # Remove ENI
+        commands = [
+            {
+                "name": "eni_id",
+                "op": "remove",
+                "type": "SAI_OBJECT_TYPE_ENI",
+            },
+        ]
+
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values remove =======")
+        pprint(results)
+
+        assert all( [result == 0 for result in results]), "SAI_OBJECT_TYPE_ENI Remove error"
+
+        # Remove VNET
+        commands = [
+            {
+                "name": "vnet",
+                "op": "remove",
+                "type": "SAI_OBJECT_TYPE_VNET"
+            },
+        ]
+
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values remove =======")
+        pprint(results)
+
+        assert all( [result == 0 for result in results]), "SAI_OBJECT_TYPE_VNET Remove error"
+        
+        commands = [
+            {
+                "name": "ore",
+                "op": "remove",
+                "type": "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY",
+            },
+        ]
+        results = [*dpu.process_commands(commands)]
+        print("\n======= SAI commands RETURN values create =======")
+        pprint(results)
+
+        assert all( [result == 0 for result in results]), "SAI_OBJECT_TYPE_OUTBOUND_ROUTING_ENTRY Remove error"
